fix: hybrid float32/float64 for numerical stability
- float32 for transport operators (safe, fast)
- float64 for matrix inversion (numerically sensitive)
- float64 for curvature tensor computation

+++ experiments/curvature/compute.py
@@ -202,18 +202,16 @@ def compute_curvature(
     n_standpoint = len(LAYER_NAMES)
 
     # Level 2 regularization: Tikhonov damping before inversion
-    U_exp_reg = U_exp + EPSILON_INV * np.eye(d_model, dtype=np.float32)
-
-    # Force float32 for speed
-    U_exp_reg = U_exp_reg.astype(np.float32)
+    # Promote to float64 for inversion (numerically sensitive)
+    U_exp_reg = (U_exp + EPSILON_INV * np.eye(d_model, dtype=np.float32)).astype(np.float64)
 
     try:
         U_exp_inv = np.linalg.inv(U_exp_reg)
     except np.linalg.LinAlgError:
         U_exp_inv = np.linalg.pinv(U_exp_reg)
 
-    # Curvature tensor
-    F = U_12 @ U_23 @ U_exp_inv
+    # Curvature tensor (promote to float64 for precision)
+    F = (U_12.astype(np.float64) @ U_23.astype(np.float64)) @ U_exp_inv
 
     # Block-specific norms
     block_norms: Dict[str, float] = {}
